// src/app/models/client.model.ts
export class Client{
    id: number;
    nCin: string; 
    nomComplet: string;
    dateNaissance: Date;
    lieuNaissance: string;
    prenomMere: string;
    prenomPere: string;
    satatMatrimoniale: string;
    nomConjoint: string;
    dateMariage: Date;
    lieuMariage: string;
    adresseHabitat: string;
    adresseCin: string;
    numTel: string;
    email: string;
    profession: string;
}